# perf: no refresh loader on new bookmark