### before

```
[dev] POST /bookmarks/new?_data=routes%2Fbookmarks%2Fnew 200 - - 19.017 ms
[dev] GET /videos/26?_data=root 200 - - 9.534 ms
[dev] GET /videos/26?_data=routes%2Fvideos%2F%24id 200 - - 11.649 ms
```

### after

```
[dev] POST /bookmarks/new?_data=routes%2Fbookmarks%2Fnew 200 - - 12.459 ms
```